added the package dependency
